pyproject.toml
[project]
name = "smbs"
version = "0.1.0"
description = "Speech Model Benchmarking Suite — encode, train, evaluate"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyannote-audio>=4.0.3",
    "torch>=2.8.0",
    "polars-lts-cpu>=1.8.0",
    "tqdm>=4.67.1",
    "ten-vad>=1.0.6.8",
    "plotly>=6.5.2",
    "nbformat>=5.10.4",
    "ipykernel>=7.1.0",
    "spidr>=0.1.3",
    "webdataset>=1.0.2",
    "transformers[torch]>=4.57.6",
    "kaleido>=1.2.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]

[project.scripts]
smbs = "smbs.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
